chore(breadcrumbs): address code reviw comments
