+++ projects/demo/src/app/breadcrumbs/breadcrumbs-routing.demo.html
@@ -22,7 +22,7 @@
         <a clrVerticalNavLink routerLink="./tools" routerLinkActive="active">Tools</a>
       </clr-vertical-nav>
       <div class="content-area breadcrumbs-content-area">
-        <clr-breadcrumbs [items]="menuItems"></clr-breadcrumbs>
+        <clr-breadcrumbs [items]="menuItems | async"></clr-breadcrumbs>
         <router-outlet></router-outlet>
       </div>
     </div>
